runtime: Move both quote and reset code in a single file
